Added a doc test.
Mostly just wanted to have an example of one in here.

BFD-1071
