+++ apps/utils/db-query-checker/Cargo.toml
@@ -3,8 +3,6 @@ name = "db-query-checker"
 version = "0.1.0"
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 
 # Easier idiomatic error handling for applications.
